Refactor(Dashboard): Moved pages management to Layout
