+++ src/styles/ManageContractPage.css
@@ -1,9 +1,5 @@
 /* src/styles/ManageContractPage.css */
-.manage-contract-container {
-    padding: 2rem;
-    max-width: 900px;
-    margin: 0 auto;
-}
+
 .management-section {
     background: #fff;
     padding: 2rem;
@@ -37,4 +33,73 @@
     border: 1px solid #ccc;
     flex-grow: 1; /* Растягиваем селект */
 }
-/* ... добавьте стили для кнопок .btn-secondary, .btn-tertiary ... */
+.services-list li {
+    display: flex; /* Включаем Flexbox для элемента списка */
+    justify-content: space-between; /* Распределяет элементы по краям */
+    align-items: center; /* Выравнивает по вертикали */
+    padding: 0.75rem 0;
+    border-bottom: 1px solid #eee;
+}
+
+.services-list li span {
+}
+
+
+.modal-services li {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0.75rem 0;
+    border-bottom: 1px solid #eee;
+}
+
+.accordion {
+    border: 1px solid #e0e0e0;
+    border-radius: 8px;
+    /*overflow: hidden;*/
+}
+.accordion-item {
+    border-bottom: 1px solid #e0e0e0;
+}
+.accordion-item:first-child .accordion-header {
+    border-top-left-radius: 8px;
+    border-top-right-radius: 8px;
+}
+
+.accordion-item:last-child {
+    /* Если последний элемент раскрыт, скругляем его контент */
+    border-bottom-left-radius: 8px;
+    border-bottom-right-radius: 8px;
+}
+.accordion-item:last-child .accordion-content {
+    border-bottom-left-radius: 8px;
+    border-bottom-right-radius: 8px;
+}
+
+/* Если последний элемент свернут, скругляем его заголовок */
+.accordion-item:last-child.accordion-item:not(:has(.accordion-content)) .accordion-header {
+    border-bottom-left-radius: 8px;
+    border-bottom-right-radius: 8px;
+}
+.accordion-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 1rem 1.5rem;
+    cursor: pointer;
+    background-color: #f9f9f9;
+}
+.accordion-content {
+    padding: 1.5rem;
+    background-color: #fff;
+    border-top: 1px solid #e0e0e0;
+}
+
+/* Добавим стиль для итоговой стоимости */
+.total-fee-section {
+    margin-top: 2rem;
+    padding-top: 1.5rem;
+    border-top: 2px solid #333;
+    text-align: right;
+    font-size: 1.2rem;
+}
