--- src/styles/ManageServicesPage.css
@@ -1,83 +1,0 @@
-/* src/styles/ManageServicesPage.css */
-
-
-
-
-.main-row {
-    cursor: pointer;
-    user-select: none; /* Чтобы текст не выделялся при клике */
-}
-
-.expand-icon {
-    text-align: center;
-    transition: transform 0.2s;
-}
-
-.details-row td {
-    padding: 0;
-    border-top: none; /* Убираем лишнюю границу */
-}
-
-.details-content {
-    background-color: #fafafa;
-    padding: 1.5rem 2rem;
-    animation: fadeIn 0.3s;
-}
-
-.details-content h4 {
-    margin-top: 0;
-}
-
-.services-list {
-    list-style: none;
-    padding: 0;
-    margin-bottom: 1.5rem;
-}
-
-.services-list li {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0.75rem 0;
-    border-bottom: 1px solid #eee;
-}
-
-.btn-disconnect, .btn-connect {
-    background: none;
-    border: 1px solid #ccc;
-    padding: 5px 10px;
-    border-radius: 5px;
-    cursor: pointer;
-}
-.btn-disconnect { color: #dc3545; border-color: #dc3545; }
-.btn-connect { color: #28a745; border-color: #28a745; }
-
-
-.btn-add-service {
-}
-
-.disconnected-services-list {
-    background-color: #fffbe6; /* Светло-желтый фон */
-    border: 1px solid #ffe58f;
-    border-radius: 8px;
-    padding: 1rem;
-    margin: 1rem 0;
-}
-.disconnected-services-list li {
-    margin-bottom: 0.5rem;
-}
-
-.modal-actions {
-    display: flex;
-    justify-content: flex-end;
-    gap: 1rem;
-    margin-top: 2rem;
-    border-top: 1px solid #eee;
-    padding-top: 1rem;
-}
-
-
-@keyframes fadeIn {
-    from { opacity: 0; }
-    to { opacity: 1; }
-}
